move the ResourceNameToPlural to mgmt namespace
